Add 'copy link' tools to the directory context menu
Edited generate_registry_file.py to add full and partial link copying
to the context menu for directories as well as files. Also added the
springboard logo to the context menu. Some keys in the registry were
renamed to clarify their purpose and to show that they are obviously
springboard tools created by this script.