Handles error when upissuing an open file or released draft
Catches the error thrown when trying to rename an open file and displays
an error message to the user requesting they close the file and try
again. Previously the script would exit before reaching an input command
so the user received no feedback on why the script failed.
Also checks to see whether a draft file has been released when upissuing
as draft, and throws an error if so. I.e. when trying to
"upissue - Draft" a revA2 doc after revA has been released the script
 will notify the user that the draft has been released and exit.