Fix path problems found when testing on other machines
Replaces %1 with %L in the registry edit. This seems to solve the
problem seen when running the code on other machines. When using %1 the
file path was being shortened:
"C:\Users\Someones Name\Dropbox (Work)\" was shorted to:
"C:Users\SOMEO~1\DROPB~1\"
Using %L seems to have fixed this. Although why the problem occurred in
the first place is a mystery!